tests

+++ src/HotChocolate/Core/test/Types.Scalars.Tests/SignedByteTypeTests.cs
@@ -124,8 +124,8 @@ public class SignedByteTypeTests : ScalarTypeTestBase
     }
 
     [Theory]
-    [InlineData(0, 0)]
-    [InlineData(1, 1)]
+    [InlineData(0, (sbyte)0)]
+    [InlineData(1, (sbyte)1)]
     [InlineData(sbyte.MaxValue, sbyte.MaxValue)]
     [InlineData(sbyte.MinValue, sbyte.MinValue)]
     [InlineData(null, null)]
