+++ src/HotChocolate/Core/test/Types.Scalars.Tests/UnsignedIntTypeTests.cs
@@ -124,8 +124,8 @@ public class UnsignedIntTypeTests : ScalarTypeTestBase
     }
 
     [Theory]
-    [InlineData(0, 0)]
-    [InlineData(1, 1)]
+    [InlineData(0, (uint)0)]
+    [InlineData(1, (uint)1)]
     [InlineData(uint.MaxValue, uint.MaxValue)]
     [InlineData(uint.MinValue, uint.MinValue)]
     [InlineData(null, null)]
