+++ src/HotChocolate/Core/test/Types.Scalars.Tests/UnsignedShortTypeTests .cs
@@ -124,8 +124,8 @@ public class UnsignedShortTypeTests : ScalarTypeTestBase
     }
 
     [Theory]
-    [InlineData(0, 0)]
-    [InlineData(1, 1)]
+    [InlineData(0, (ushort)0)]
+    [InlineData(1, (ushort)1)]
     [InlineData(ushort.MaxValue, ushort.MaxValue)]
     [InlineData(ushort.MinValue, ushort.MinValue)]
     [InlineData(null, null)]
